Move inline assembly stuff to a separate syscall file
Also change the main function return type to an unsigned byte

+++ include/process.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
+#include "types.hpp"
+
 namespace brb
 {
-	void exit(int exit_code);
+	void exit(u8 exit_code);
 }
